完成院校部分的分页，根据用户点击学校详情使用幻灯片实现智能推荐，修改后台幻灯片的删除部分，使得删除的同时删除磁盘文件，修复院校搜索的bug

+++ sais-entity/src/main/java/com/sais/saisentity/User.java
@@ -23,9 +23,9 @@ public class User implements Serializable {
     private int follows_num=0;
     private int fans_num=0;
     private String phone="";
-    private float gpa=0;
+    private Double gpa=0.0;
     private int sat=0;
-    private float ielts=0;
-    private float toefl=0;
+    private Double ielts=0.0;
+    private Double toefl=0.0;
     private String recommend;
 }
